requirements.txt
python-i18n==0.3.9
PyYAML==5.4.1
